Add files via upload
known_primes is a text file that contains over the first million primes, all primes up to a bit over 18 million.  having lots of pre-calculated primes can be useful.
the python program prime_discovery.py will continue to extend this known_primes file as long as you are willing to run it.